Transition to .upstream: Use in taggedRef and canonicalRef
Start transitioning from .our uses to .upstream.  First in the simplest
cases: taggedRef.Tag() and canonicalRef.Digest() are values in principle
unaffected by the name canonicalization, so this should be an obviously
correct change which does not change behavior.

Starting with this one to demostrate the principle of moving step by
step.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>